backend/pyproject.toml
[project]
name = "knowledge.curator"
version = "1.0.0.dev0"
description = "A knowledge curation tool built on Plone."
readme = "README.md"
license = "GPL-2.0-only"
requires-python = ">=3.10"
authors = [
    { name = "Darren Lund", email = "darren.lund@gauntletai.com" },
]
keywords = [
    "CMS",
    "Plone",
    "Python",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Plone","Framework :: Plone :: 6.1",
    "Framework :: Plone :: Addon",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "Products.CMFPlone==6.1.2",
    "plone.api",
    "plone.restapi",
    "plone.volto",
    "z3c.jbot",
    "plone.app.dexterity",
    "plone.app.relationfield",
    "plone.app.lockingbehavior",
    "plone.schema",
    "sentence-transformers",
    "qdrant-client",
    "redis",
    "celery",
    "reportlab",
]

[project.optional-dependencies]
test = [
    "horse-with-no-namespace",
    "plone.app.testing",
    "plone.testing",
    "plone.app.robotframework",
    "plone.restapi[test]",
    "pytest",
    "pytest-cov",
    "pytest-plone>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/GitHub/knowledge-curator"
PyPI = "https://pypi.org/project/knowledge.curator"
Source = "https://github.com/GitHub/knowledge-curator"
Tracker = "https://github.com/GitHub/knowledge-curator/issues"


[project.entry-points."plone.autoinclude.plugin"]
target = "plone"

[tool.uv]
managed = false


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
strict-naming = true

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
]

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge"]

[tool.towncrier]
directory = "news/"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
title_format = "## {version} ({project_date})"
template = "news/.changelog_template.jinja"
issue_format = "[#{issue}](https://github.com/GitHub/knowledge-curator/issues/{issue})"
underlines = ["", "", ""]

[[tool.towncrier.type]]
directory = "breaking"
name = "Breaking changes:"
showcontent = true

[[tool.towncrier.type]]
directory = "feature"
name = "New features:"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bug fixes:"
showcontent = true

[[tool.towncrier.type]]
directory = "internal"
name = "Internal:"
showcontent = true

[[tool.towncrier.type]]
directory = "documentation"
name = "Documentation:"
showcontent = true

[[tool.towncrier.type]]
directory = "tests"
name = "Tests"
showcontent = true

[tool.ruff]
target-version = "py312"
line-length = 88
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
]
lint.ignore = [
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.isort]
case-sensitive = false
no-sections = true
force-single-line = true
from-first = true
lines-after-imports = 2
lines-between-types = 1
order-by-type = false

[tool.ruff.format]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501", "RUF001", "S101"]

[tool.check-manifest]
ignore = [
    ".editorconfig",
    ".flake8",
    "dependabot.yml",
    "mx.ini",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source_pkgs = ["knowledge.curator", "tests"]
branch = true
parallel = true
omit = [
  "src/knowledge/curator/locales/*.py",
]

[tool.pyright]
# Configuration for Plone development environment
venvPath = "."
venv = "venv"
pythonPlatform = "All"

# Suppress warnings only for specific external packages that may not be available during development
reportMissingTypeStubs = "none"

# Specific packages to ignore import warnings for
extraPaths = [
    "src"
]

# Allow undefined variables as warnings rather than errors in test files
reportUndefinedVariable = "warning"

# Suppress specific warnings for external packages
reportOptionalCall = "none"
reportOptionalMemberAccess = "none"

# Ignore specific external package patterns
reportMissingModuleSource = "none"